solicitor/requirements.txt
python-dotenv>=1.0
pyyaml>=6.0
requests>=2.31
python-dateutil>=2.9
setlist-fm-client>=0.6
spotipy>=2.24
apple-music-python>=0.2
musicbrainzngs>=0.7
